Update linux-static-proxy-configuration.md
In the MDE Linux script, the configuration is set to read the localconfig.json file for Azure Arc, and we have confirmed that this is the current specification.

I have investigated this behavior in our test tenant and also from the customer. Customer requested to write this to docs as a current behavior.